Fix Windows tests

+++ test/libsyncengine/jobs/network/testnetworkjobs.cpp
@@ -1234,7 +1234,7 @@ void TestNetworkJobs::testDriveUploadSessionAsynchronousAborted() {
     JobManager::instance()->queueAsyncJob(driveUploadSessionJob);
 
     int counter = 0;
-    while (driveUploadSessionJob->state() <= DriveUploadSession::StateStartUploadSession) {
+    while (static_cast<int>(driveUploadSessionJob->state()) <= static_cast<int>(DriveUploadSession::StateStartUploadSession)) {
         Utility::msleep(10);
         CPPUNIT_ASSERT_LESS(500, ++counter); // Wait at most 5sec
     }
